syncing land works, stable state
all it took was to move the lands out of the drop zone, and giving them each a network identity.

I believe I had a similar issue when it came to the deck. I just moved it to player manager.

There are certain conditions required for objects to be synced, whether that be where they are stored locally or in terms of parent-child.