🎉 MagicMoney overhauled and fully functional
Magic and Money have been moved to Player (as opposed to their own scripts), and correctly count up as game progresses.

It's getting kind of open-ended now in terms of what pieces of the game I move to for the next implementation.

For now though, I will look at the turn counter. Then move to fully functional combat, which will be a beast of its own.