lobby creation cont.
(ChatGPT) swapped the lobby player for the in-game player

Works as intended, flows from menu > lobby > game like we just added a step.

Interestingly i got a bunch of null refs, because of the nature of swapping the player prefabs. The fix (which i figured on my own, thank you) was as simple as changing that gameManager.FullLobby() to something the joiner called only. The function was commented something completely wrong for a while. Take a look.