Start of EOS
So in order to have the game be playable across networks, you need a third party service. I went with Epic Games because I don't need to pay for server up keep or have users solve the da Vinci code (port relaying) to be able to play the game with me.

This is just the beginning I got the package in and configured. Here it is if you are curious: https://github.com/EOS-Contrib/eos_plugin_for_unity