copied over EOSTransport stuff.
goddamn that's a lot of files. It says 'Commit 1725 files to v2,' at the bottom. Let's hope it actually commits....

So the story behind this one is that the eos_plugin_for_unity only has the EOSManager Component, which is weird and stupid. I had to look around and find this repo down below for literally everything else (EosTransport, the SDK, etc.).

https://github.com/FakeByte/EpicOnlineTransport?tab=readme-ov-file