EOS rollback
EOSBootstrap was a function GPT told me to use.

Apparently my credentials for using the DevAuthTool are incorrect. This results in EOS never being initialized and the server not starting. This has been the on and off issue for a while.

EOS is a service the Epic provides that would allow play across different LANS, aka different wifi networks. Mirror on it's own only allows for same LAN play.

This has proven to be a major headache, so I'm going to split off and return later. I don't need a port relay or server to have a working game.